fix: Disable ALPN protocol selection when Reality is active
Adds a dependency to the ALPN protocol selection option in the Ray server
configuration. This option will now only be visible when TLS is enabled
and Reality is disabled.

This change prevents users from configuring ALPN protocols when Reality
is in use, avoiding potential conflicts or misconfigurations as Reality
might manage these aspects internally or not support custom ALPN settings.